fixed a logging bug

+++ HW2/hw2.py
@@ -316,9 +316,10 @@ def clear_log_file() -> None:
     """
     Clears the hw2_logfile.log file.
     """
-    with open(LOG_FILE_PATH, "w") as logfile:
-        logfile.seek(0)
-        logfile.write("")
+    if LOG and os.path.exists(LOG_FILE_PATH):
+        with open(LOG_FILE_PATH, "w") as logfile:
+            logfile.seek(0)
+            logfile.write("")
 
     return None
 
@@ -333,6 +334,9 @@ def logger_init() -> logging.Logger:
         at the INFO level and above. It adds a FileHandler to the logger, directing log messages to a
         specified log file.
     """
+    log_directory = os.path.join(os.getcwd(), "logs")
+    os.makedirs(log_directory, exist_ok=True)
+  
     try:
         logger: logging.Logger = logging.getLogger(LOGGER_NAME)
         logger.setLevel(logging.DEBUG)
